Move pkg/kubectl/version to staging
