+++ cmd/kubectl/BUILD
@@ -3,7 +3,7 @@ load(
     "go_binary",
     "go_library",
 )
-load("//pkg/kubectl/version:def.bzl", "version_x_defs")
+load("//staging/src/k8s.io/kubectl/pkg/version:def.bzl", "version_x_defs")
 
 go_binary(
     name = "kubectl",
