+++ hack/print-workspace-status.sh
@@ -30,7 +30,7 @@ kube::version::get_version_vars
 # Stamped rules will be retriggered by changes to stable-status.txt, but not by
 # changes to volatile-status.txt.
 # IMPORTANT: the camelCase vars should match the lists in hack/lib/version.sh
-# and pkg/version/def.bzl.
+# and staging/src/k8s.io/kubectl/pkg/version/def.bzl.
 cat <<EOF
 STABLE_BUILD_GIT_COMMIT ${KUBE_GIT_COMMIT-}
 STABLE_BUILD_SCM_STATUS ${KUBE_GIT_TREE_STATE-}
