+++ pkg/kubectl/BUILD
@@ -91,7 +91,6 @@ filegroup(
         "//pkg/kubectl/polymorphichelpers:all-srcs",
         "//pkg/kubectl/proxy:all-srcs",
         "//pkg/kubectl/util:all-srcs",
-        "//pkg/kubectl/version:all-srcs",
     ],
     tags = ["automanaged"],
 )
