+++ pkg/kubectl/cmd/util/kubectl_match_version.go
@@ -29,7 +29,7 @@ import (
 	"k8s.io/kubectl/pkg/scheme"
 
 	"k8s.io/cli-runtime/pkg/genericclioptions"
-	"k8s.io/kubernetes/pkg/kubectl/version"
+	"k8s.io/kubectl/pkg/version"
 )
 
 const (
