+++ staging/src/k8s.io/kubectl/BUILD
@@ -16,6 +16,7 @@ filegroup(
         "//staging/src/k8s.io/kubectl/pkg/scheme:all-srcs",
         "//staging/src/k8s.io/kubectl/pkg/util:all-srcs",
         "//staging/src/k8s.io/kubectl/pkg/validation:all-srcs",
+        "//staging/src/k8s.io/kubectl/pkg/version:all-srcs",
     ],
     tags = ["automanaged"],
     visibility = ["//visibility:public"],
